+++ docs/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project does not adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 This project uses MCVERSION-MAJORMOD.MAJORAPI.MINOR.PATCH.
 
+## [Unreleased]
+### Added
+- Config options for certain affix settings, currently there are settings for the Horde and Lively affixes
+
 ## [0.3] - 2018.11.23
 ### Added
 - Plagued affix - Infects anything nearby with Plague, which adds a Wither effect to the target and the target becomes a new Plague host